restructure assignment project
