--- Data/Client.cs
@@ -1,10 +1,0 @@
-namespace CareKickoff.Data;
-
-public class Client
-{
-	public String FirstName { get; set; }
-    public String LastName { get; set; }
-    public DateTime BirthDate { get; set; }
-    public String Gender { get; set; }
-    public int NativeId { get; set; }
-}
