+++ Models/Client.cs
@@ -1,0 +1,10 @@
+namespace CareKickoff.Models;
+
+public class Client
+{
+    public string FirstName { get; set; }
+    public string LastName { get; set; }
+    public DateTime BirthDate { get; set; }
+    public string Gender { get; set; }
+    public int NativeId { get; set; }
+}
