+++ Models/Employee.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CareKickoff.Data
+namespace CareKickoff.Models
 {
     public class Employee
     {
-        public String Name { get; set; }
+        public string Name { get; set; }
         public int EmployeeId { get; set; }
         public List<int> AuthorizedClients { get; set; }
     }
