+++ Models/ICareplanService.cs
@@ -1,0 +1,7 @@
+﻿namespace CareKickoff.Models
+{
+    public interface ICareplanService
+    {
+        public Task<Careplan[]> GetCarePlansForClientAsync(int clientId);
+    }
+}
