+++ Models/IClientService.cs
@@ -1,0 +1,9 @@
+﻿namespace CareKickoff.Models
+{
+    public interface IClientService
+    {
+        public Task<Client[]> GetAllClientsAsync();
+
+        public Task<Client> GetClientByIdAsync(int id);
+    }
+}
