+++ Models/IEmployeeService.cs
@@ -1,0 +1,10 @@
+﻿namespace CareKickoff.Models
+{
+    public interface IEmployeeService
+    {
+        public Task<Employee[]> GetAllEmployeesAsync();
+
+        public Task<Employee> GetEmployeeByIdAsync(int id);
+        
+    }
+}
