+++ Models/IReportService.cs
@@ -1,0 +1,7 @@
+﻿namespace CareKickoff.Models
+{
+    public interface IReportService
+    {
+        public Task<Report[]> GetReportsForClientAsync(int clientId);
+    }
+}
