+++ Pages/ClientOverview.razor
@@ -1,6 +1,7 @@
 ﻿@page "/clientoverview"
 
 @using CareKickoff.Data
+@using CareKickoff.Models
 @inject ClientService ClientsService
 @inject NavigationManager NavigationManager
