+++ Pages/ClientView.razor
@@ -2,6 +2,7 @@
 @page "/clientview/{clientId:int}"
 
 @using CareKickoff.Data
+@using CareKickoff.Models
 @inject ClientService ClientsService
 @inject ReportService ReportService
 @inject CareplanService CareplanService
