+++ Pages/EmployeeOverview.razor
@@ -1,6 +1,7 @@
 ﻿@page "/employeeoverview"
 
 @using CareKickoff.Data
+@using CareKickoff.Models
 @inject EmployeeService EmployeeService
