+++ Pages/ReportOverview.razor
@@ -1,6 +1,7 @@
 @page "/reports/{clientId:int}"
 
 @using CareKickoff.Data
+@using CareKickoff.Models
 @inject ReportService ReportService
 
 @if (reportsForClient == null)
